pyproject.toml
[project]
name = "ulearning-questiontrain-export"
version = "0.1.0"
description = "导出优学院题库训练的题目"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
]
